Renamed goal from 'build-gnur' to 'gnur-compile' to be consistent with other goals
