+++ test-packages/dataTest/pom.xml
@@ -36,7 +36,7 @@
                     <execution>
                         <id>build</id>
                         <goals>
-                            <goal>build-gnur</goal>
+                            <goal>gnur-compile</goal>
                         </goals>
                         <phase>compile</phase>
                     </execution>
