+++ test-packages/link/pom.xml
@@ -43,7 +43,7 @@
           <execution>
             <id>build</id>
             <goals>
-              <goal>build-gnur</goal>
+              <goal>gnur-compile</goal>
             </goals>
           </execution>
         </executions>
