+++ test-packages/native/pom.xml
@@ -37,7 +37,7 @@
           <execution>
             <id>compile-c-sources</id>
             <goals>
-              <goal>build-gnur</goal>
+              <goal>gnur-compile</goal>
             </goals>
             <phase>compile</phase>
           </execution>
